+++ src/test/test.py
@@ -11,12 +11,11 @@ import subprocess
 import tempfile
 import time
 import unittest
-from shutil import copy, copytree
+from shutil import copy
 from string import Template
 
 import djerba.util.constants as constants
 import djerba.util.ini_fields as ini
-from djerba.benchmark import benchmarker
 from djerba.configure import configurer, log_r_cutoff_finder
 from djerba.extract.extractor import extractor
 from djerba.extract.r_script_wrapper import r_script_wrapper
@@ -26,6 +25,7 @@ from djerba.mavis import mavis_runner
 from djerba.render.archiver import archiver
 from djerba.render.render import html_renderer, pdf_renderer
 from djerba.sequenza import sequenza_reader, SequenzaError
+from djerba.util.provenance_reader import InsufficientSampleNamesError, UnknownTumorNormalIDError
 from djerba.util.validator import config_validator, DjerbaConfigError
 
 class TestBase(unittest.TestCase):
@@ -51,12 +51,19 @@ class TestBase(unittest.TestCase):
         self.data_dir = os.path.realpath(os.path.join(test_dir, 'data'))
         # specify all non-public data paths relative to self.sup_dir
         sup_dir_var = 'DJERBA_TEST_DATA'
+        bench_dir_var = 'DJERBA_GSICAPBENCH_DATA'
         self.sup_dir = os.environ.get(sup_dir_var)
+        self.bench_dir = os.environ.get(bench_dir_var)
         if not (self.sup_dir):
             raise RuntimeError('Need to specify environment variable {0}'.format(sup_dir_var))
         elif not os.path.isdir(self.sup_dir):
             raise OSError("Supplementary directory path '{0}' is not a directory".format(self.sup_dir))
+        if not (self.bench_dir):
+            raise RuntimeError('Need to specify environment variable {0}'.format(bench_dir_var))
+        elif not os.path.isdir(self.bench_dir):
+            raise OSError("GSICAPBENCH directory path '{0}' is not a directory".format(self.bench_dir))
         self.provenance = os.path.join(self.sup_dir, 'pass01_panx_provenance.original.tsv.gz')
+        self.provenance_vnwgts = os.path.join(self.sup_dir, 'provenance_VNWGTS_0329.tsv.gz')
         self.tmp = tempfile.TemporaryDirectory(prefix='djerba_')
         self.tmp_dir = self.tmp.name
         self.bed_path = os.path.join(self.sup_dir, 'S31285117_Regions.bed')
@@ -69,6 +76,9 @@ class TestBase(unittest.TestCase):
             self.config_user,
             self.config_user_wgs_only,
             self.config_user_failed,
+            self.config_user_vnwgts,
+            self.config_user_vnwgts_broken_1,
+            self.config_user_vnwgts_broken_2,
             self.config_user_wgs_only_failed,
             self.config_full,
             self.config_full_wgs_only,
@@ -93,6 +103,9 @@ class TestBase(unittest.TestCase):
                 'config_user.ini',
                 'config_user_wgs_only.ini',
                 'config_user_failed.ini',
+                'config_user_vnwgts.ini',
+                'config_user_vnwgts_broken_1.ini',
+                'config_user_vnwgts_broken_2.ini',
                 'config_user_wgs_only_failed.ini',
                 'config_full.ini',
                 'config_full_wgs_only.ini',
@@ -118,65 +131,16 @@ class TestArchive(TestBase):
         # contents of file are dependent on local paths
         with open(archive_path) as archive_file:
             data = json.loads(archive_file.read())
-        self.assertEqual(len(data['report']), 19)
+        self.assertEqual(len(data['report']), 20)
         self.assertEqual(len(data['supplementary']['config']), 3)
-
-class TestBenchmark(TestBase):
-
-    class mock_args_compare:
-        """Use instead of argparse to store params for testing"""
-
-        def __init__(self, report_dirs, compare_all=False):
-            self.subparser_name = constants.COMPARE
-            self.report_dir = report_dirs
-            self.compare_all = compare_all
-            # logging
-            self.log_path = None
-            self.debug = False
-            self.verbose = False
-            self.quiet = True
-
-    class mock_args_report:
-        """Use instead of argparse to store params for testing"""
-
-        def __init__(self, input_dir, output_dir):
-            self.subparser_name = constants.REPORT
-            self.input_dir = input_dir
-            self.output_dir = output_dir
-            self.dry_run = False
-            # logging
-            self.log_path = None
-            self.debug = False
-            self.verbose = False
-            self.quiet = True
-
-    def test_benchmark(self):
-        out_dir = self.tmp_dir
-        input_dir = os.path.join(self.sup_dir, 'benchmark')
-        report_dir = os.path.join(out_dir, 'report')
-        os.mkdir(report_dir)
-        report_args = self.mock_args_report(input_dir, report_dir)
-        self.assertTrue(benchmarker(report_args).run())
-        report_1a = os.path.join(report_dir, 'GSICAPBENCH_1219')
-        report_1b = os.path.join(report_dir, 'GSICAPBENCH_1219.copy')
-        copytree(report_1a, report_1b) # make a copy to test identical inputs
-        report_2 = os.path.join(report_dir, 'GSICAPBENCH_1232')
-        compare_args_1 = self.mock_args_compare([report_1a, report_1b])
-        self.assertTrue(benchmarker(compare_args_1).run())
-        compare_args_2 = self.mock_args_compare([report_1a, report_2])
-        self.assertFalse(benchmarker(compare_args_2).run())
-        compare_args_3 = self.mock_args_compare([report_1a, report_1b], compare_all=True)
-        self.assertTrue(benchmarker(compare_args_3).run())
-        compare_args_4 = self.mock_args_compare([report_1a, report_2], compare_all=True)
-        self.assertFalse(benchmarker(compare_args_4).run())
 
 class TestConfigure(TestBase):
 
-    def run_config_test(self, user_config, wgs_only, failed, expected_lines):
+    def run_config_test(self, user_config, wgs_only, failed, expected_lines, provenance):
         config = configparser.ConfigParser()
         config.read(self.default_ini)
         config.read(user_config)
-        config['settings']['provenance'] = self.provenance
+        config['settings']['provenance'] = provenance
         test_configurer = configurer(config, wgs_only, failed)
         out_dir = self.tmp_dir
         out_path = os.path.join(out_dir, 'config_test_output.ini')
@@ -187,18 +151,38 @@ class TestConfigure(TestBase):
             lines = len(out_file.readlines())
         self.assertEqual(lines, expected_lines) # unlike archive, configParser puts a blank line at the end of the file
 
+    def run_config_broken(self, user_config, provenance):
+        # no assertions; run is intended to fail
+        config = configparser.ConfigParser()
+        config.read(self.default_ini)
+        config.read(user_config)
+        config['settings']['provenance'] = provenance
+        test_configurer = configurer(config, wgs_only=False, failed=False, log_level=logging.CRITICAL)
+        out_dir = self.tmp_dir
+        out_path = os.path.join(out_dir, 'config_test_output.ini')
+        test_configurer.run(out_path)
+
     def test_default(self):
-        self.run_config_test(self.config_user, False, False, 52)
+        self.run_config_test(self.config_user, False, False, 55, self.provenance)
 
     def test_default_fail(self):
-        self.run_config_test(self.config_user_failed, False, True, 42)
+        self.run_config_test(self.config_user_failed, False, True, 45, self.provenance)
 
     def test_wgs_only(self):
-        self.run_config_test(self.config_user_wgs_only, True, False, 50)
+        self.run_config_test(self.config_user_wgs_only, True, False, 53, self.provenance)
 
     def test_wgs_only_fail(self):
-        self.run_config_test(self.config_user_wgs_only_failed, True, True, 42)
+        self.run_config_test(self.config_user_wgs_only_failed, True, True, 45, self.provenance)
 
+    def test_vnwgts(self):
+        self.run_config_test(self.config_user_vnwgts, False, False, 55, self.provenance_vnwgts)
+
+    def test_vnwgts_broken(self):
+        # test failure modes of sample input
+        with self.assertRaises(InsufficientSampleNamesError):
+            self.run_config_broken(self.config_user_vnwgts_broken_1, self.provenance_vnwgts)
+        with self.assertRaises(UnknownTumorNormalIDError):
+            self.run_config_broken(self.config_user_vnwgts_broken_2, self.provenance_vnwgts)
 
 class TestCutoffFinder(TestBase):
 
@@ -242,11 +226,11 @@ class TestExtractor(TestBase):
     STATIC_MD5_FAILED = {
         'data_clinical.txt': 'ec0868407eeaf100dbbbdbeaed6f1774',
         'genomic_summary.txt': '5a2f6e61fdf0f109ac3d1bcc4bb3ca71',
-        'djerba_report.json': '8eca62a9ffd80310f8298161846d0912'
+        'djerba_report.json': 'b114649420104ec7b1b2dd773cb9f44c'
     }
     VARYING_OUTPUT = [
-        'tmb.jpeg',
-        'vaf.jpeg',
+        'tmb.svg',
+        'vaf.svg',
         'djerba_report.json'
     ]
 
@@ -262,6 +246,8 @@ class TestExtractor(TestBase):
         # do not check supplementary data
         del data_found['supplementary']
         del data_expected['supplementary']
+        # replace djerba version with a placeholder
+        data_found['report']['djerba_version'] = 'PLACEHOLDER'
         self.maxDiff = None
         self.assertEqual(data_found, data_expected)
 
@@ -361,6 +347,10 @@ class TestLister(TestBase):
             self.study = project
             self.provenance = provenance
             self.wgs_only = False
+            # sample names
+            self.wgn = None
+            self.wgt = None
+            self.wtt = None
             # logging
             self.log_path = None
             self.debug = False
@@ -469,6 +459,10 @@ class TestMavis(TestBase):
             self.work_dir = work_dir
             self.donor = donor
             self.study = study
+            # sample names
+            self.wgn = None
+            self.wgt = None
+            self.wtt = None
             # logging
             self.log_path = None
             self.debug = False
@@ -539,15 +533,15 @@ class TestRender(TestBase):
         args_path = os.path.join(self.sup_dir, 'report_json', 'WGTS', 'djerba_report.json')
         out_path = os.path.join(self.tmp_dir, 'djerba_test_wgts.html')
         html_renderer().run(args_path, out_path, False)
-        self.check_report(out_path, '6eed39e0f315b3687a18a1a87e39c5b4')
+        self.check_report(out_path, '7d9a02918f392c07975ce6fbcf1c4072')
         args_path = os.path.join(self.sup_dir, 'report_json', 'WGS_only', 'djerba_report.json')
         out_path = os.path.join(self.tmp_dir, 'djerba_test_wgs_only.html')
         html_renderer().run(args_path, out_path, False)
-        self.check_report(out_path, '8977971426c0568783fab8b060d1351f')
+        self.check_report(out_path, 'd120aba19f4653f25a7a501167487f60')
         args_path = os.path.join(self.sup_dir, 'report_json', 'failed', 'djerba_report.json')
         out_path = os.path.join(self.tmp_dir, 'djerba_test_failed.html')
         html_renderer().run(args_path, out_path, False)
-        self.check_report(out_path, '2458deb2afa230d6886dbf82d49bef75')
+        self.check_report(out_path, 'f421c403f90f8a7335c292519f9b7c69')
 
     def test_pdf(self):
         in_path = os.path.join(self.sup_dir, 'djerba_test.html')